requirements.txt
# Data analysis
matplotlib

# Database
sqlalchemy>=2.0.0
psycopg2-binary
alembic

# Configuration
python-dotenv

# HTTP service and OCR
fastapi
uvicorn
requests
# Optional OCR (PaddleOCR-VL)
paddlepaddle-gpu  # or paddlepaddle for CPU-only
paddleocr
